Refactored login and signup
